requirements.txt
tabula-py==2.2.0
pandas==1.0.3
PyPDF2==1.26.0
urllib3==1.25.8
sklearn
SQLAlchemy==1.4.6